ref(issues): Refactor group merged components
Move to shared folder as these are reusable between org and project
issue details.

Depends on https://github.com/getsentry/sentry/pull/11519
