+++ src/sentry/static/sentry/app/routes.jsx
@@ -6,7 +6,7 @@ import CreateProject from 'app/views/onboarding/createProject';
 import ProjectGroupDetails from 'app/views/groupDetails/project/index';
 import ProjectGroupEvents from 'app/views/groupDetails/project/groupEvents';
 import ProjectGroupEventDetails from 'app/views/groupDetails/project/groupEventDetails';
-import ProjectGroupMergedView from 'app/views/groupDetails/project/groupMerged';
+import ProjectGroupMergedView from 'app/views/groupDetails/shared/groupMerged';
 import ProjectGroupSimilarView from 'app/views/groupDetails/project/groupSimilar';
 import ProjectGroupTagValues from 'app/views/groupDetails/project/groupTagValues';
 import ProjectGroupTags from 'app/views/groupDetails/project/groupTags';
