+++ src/sentry/static/sentry/app/views/groupDetails/shared/groupMerged/mergedList.jsx
@@ -29,7 +29,7 @@ class MergedList extends React.Component {
   };
 
   render() {
-    let {items, pageLinks, onToggleCollapse, onUnmerge, ...otherProps} = this.props;
+    let {items, pageLinks, onToggleCollapse, onUnmerge} = this.props;
     let itemsWithLatestEvent = items.filter(({latestEvent}) => !!latestEvent);
     let hasResults = itemsWithLatestEvent.length > 0;
 
@@ -44,11 +44,7 @@ class MergedList extends React.Component {
           <QueryCount count={itemsWithLatestEvent.length} />
         </h2>
 
-        <MergedToolbar
-          {...otherProps}
-          onToggleCollapse={onToggleCollapse}
-          onUnmerge={onUnmerge}
-        />
+        <MergedToolbar onToggleCollapse={onToggleCollapse} onUnmerge={onUnmerge} />
 
         <MergedItems>
           {itemsWithLatestEvent.map(({id, latestEvent}) => (
